Refactor TorchDataset such that its __getitem__ method returns a BatchData instance instead of a dictionary
