+++ modelforge/dataset/dataset.py
@@ -1,11 +1,12 @@
 import os
 from dataclasses import dataclass
-from typing import TYPE_CHECKING, Dict, List, Literal, Optional, Union
+from typing import TYPE_CHECKING, Dict, List, Literal, Optional, Union, NamedTuple
 
 import numpy as np
 import pytorch_lightning as pl
 import torch
 from loguru import logger as log
+from openff.units import Quantity
 from torch.utils.data import DataLoader
 
 from modelforge.dataset.utils import RandomRecordSplittingStrategy, SplittingStrategy
@@ -23,6 +24,158 @@ class DatasetStatistics:
     atomic_self_energies_removed: bool = False
 
 
+@dataclass(frozen=False)
+class Metadata:
+    """
+    A NamedTuple to structure the inputs for neural network potentials.
+
+    Parameters
+    ----------
+    """
+
+    E: torch.Tensor
+    atomic_subsystem_counts: torch.Tensor
+    atomic_subsystem_indices_referencing_dataset: torch.Tensor
+    number_of_atoms: int
+    F: torch.Tensor = torch.tensor([], dtype=torch.float32)
+
+    def to(
+            self, device: Optional[torch.device] = None, dtype: Optional[torch.dtype] = None
+    ):
+        """Move all tensors in this instance to the specified device."""
+        if device:
+            self.E = self.E.to(device)
+            self.F = self.F.to(device)
+            self.atomic_subsystem_counts = self.atomic_subsystem_counts.to(device)
+            self.atomic_subsystem_indices_referencing_dataset = (
+                self.atomic_subsystem_indices_referencing_dataset.to(device)
+            )
+        if dtype:
+            self.E = self.E.to(dtype)
+            self.F = self.F.to(dtype)
+        return self
+
+@dataclass
+class NNPInput:
+    """
+    A dataclass to structure the inputs for neural network potentials.
+
+    Attributes
+    ----------
+    atomic_numbers : torch.Tensor
+        A 1D tensor containing atomic numbers for each atom in the system(s).
+        Shape: [num_atoms], where `num_atoms` is the total number of atoms across all systems.
+    positions : torch.Tensor
+        A 2D tensor of shape [num_atoms, 3], representing the XYZ coordinates of each atom.
+    atomic_subsystem_indices : torch.Tensor
+        A 1D tensor mapping each atom to its respective subsystem or molecule.
+        This allows for calculations involving multiple molecules or subsystems within the same batch.
+        Shape: [num_atoms].
+    total_charge : torch.Tensor
+        A tensor with the total charge of molecule.
+        Shape: [num_systems], where `num_systems` is the number of molecules.
+    """
+
+    atomic_numbers: torch.Tensor
+    positions: Union[torch.Tensor, Quantity]
+    atomic_subsystem_indices: torch.Tensor
+    total_charge: torch.Tensor
+
+    def to(
+            self,
+            *,
+            device: Optional[torch.device] = None,
+            dtype: Optional[torch.dtype] = None,
+    ):
+        """Move all tensors in this instance to the specified device/dtype."""
+
+        if device:
+            self.atomic_numbers = self.atomic_numbers.to(device)
+            self.positions = self.positions.to(device)
+            self.atomic_subsystem_indices = self.atomic_subsystem_indices.to(device)
+            self.total_charge = self.total_charge.to(device)
+        if dtype:
+            self.positions = self.positions.to(dtype)
+        return self
+
+    def __post_init__(self):
+        # Set dtype and convert units if necessary
+        self.atomic_numbers = self.atomic_numbers.to(torch.int32)
+        self.atomic_subsystem_indices = self.atomic_subsystem_indices.to(torch.int32)
+        self.total_charge = self.total_charge.to(torch.int32)
+
+        # Unit conversion for positions
+        if isinstance(self.positions, Quantity):
+            positions = self.positions.to(unit.nanometer).m
+            self.positions = torch.tensor(
+                positions, dtype=torch.float32, requires_grad=True
+            )
+
+        # Validate inputs
+        self._validate_inputs()
+
+    def _validate_inputs(self):
+        if self.atomic_numbers.dim() != 1:
+            raise ValueError("atomic_numbers must be a 1D tensor")
+        if self.positions.dim() != 2 or self.positions.size(1) != 3:
+            raise ValueError("positions must be a 2D tensor with shape [num_atoms, 3]")
+        if self.atomic_subsystem_indices.dim() != 1:
+            raise ValueError("atomic_subsystem_indices must be a 1D tensor")
+        if self.total_charge.dim() != 1:
+            raise ValueError("total_charge must be a 1D tensor")
+
+        # Optionally, check that the lengths match if required
+        if len(self.positions) != len(self.atomic_numbers):
+            raise ValueError(
+                "The size of atomic_numbers and the first dimension of positions must match"
+            )
+        if len(self.positions) != len(self.atomic_subsystem_indices):
+            raise ValueError(
+                "The size of atomic_subsystem_indices and the first dimension of positions must match"
+            )
+
+    def as_namedtuple(self) -> NamedTuple:
+        """Export the dataclass fields and values as a named tuple."""
+
+        import collections
+        from dataclasses import dataclass, fields
+
+        NNPInputTuple = collections.namedtuple(
+            "NNPInputTuple", [field.name for field in fields(self)]
+        )
+        return NNPInputTuple(*[getattr(self, field.name) for field in fields(self)])
+
+    def as_jax_namedtuple(self) -> NamedTuple:
+        """Export the dataclass fields and values as a named tuple.
+        Convert pytorch tensors to jax arrays."""
+
+        from dataclasses import dataclass, fields
+        import collections
+        from pytorch2jax.pytorch2jax import convert_to_jax
+
+        NNPInputTuple = collections.namedtuple(
+            "NNPInputTuple", [field.name for field in fields(self)]
+        )
+        return NNPInputTuple(
+            *[convert_to_jax(getattr(self, field.name)) for field in fields(self)]
+        )
+
+
+@dataclass
+class BatchData:
+    nnp_input: NNPInput
+    metadata: Metadata
+
+    def to(
+            self,
+            device: Optional[torch.device] = None,
+            dtype: Optional[torch.dtype] = None,
+    ):
+        self.nnp_input = self.nnp_input.to(device=device, dtype=dtype)
+        self.metadata = self.metadata.to(device=device, dtype=dtype)
+        return self
+
+
 class TorchDataset(torch.utils.data.Dataset[Dict[str, torch.Tensor]]):
     """
     Wraps a numpy dataset to make it compatible with PyTorch DataLoader.
@@ -37,8 +190,6 @@ class TorchDataset(torch.utils.data.Dataset[Dict[str, torch.Tensor]]):
         If True, converts properties to PyTorch tensors ahead of time. Default is False.
 
     """
-
-    # TODO: add support for general properties with given formats
 
     def __init__(
         self,
@@ -95,12 +246,12 @@ class TorchDataset(torch.utils.data.Dataset[Dict[str, torch.Tensor]]):
         self.number_of_records = len(dataset["atomic_subsystem_counts"])
         self.number_of_atoms = len(dataset["atomic_numbers"])
         single_atom_start_idxs_by_rec = np.concatenate(
-            [[0], np.cumsum(dataset["atomic_subsystem_counts"])]
+            [np.array([0]), np.cumsum(dataset["atomic_subsystem_counts"])]
         )
         # length: n_records + 1
 
         self.series_mol_start_idxs_by_rec = np.concatenate(
-            [[0], np.cumsum(dataset["n_confs"])]
+            [np.array([0]), np.cumsum(dataset["n_confs"])]
         )
         # length: n_records + 1
 
@@ -120,7 +271,7 @@ class TorchDataset(torch.utils.data.Dataset[Dict[str, torch.Tensor]]):
 
         self.series_atom_start_idxs_by_conf = np.concatenate(
             [
-                [0],
+                np.array([0]),
                 np.cumsum(
                     np.repeat(dataset["atomic_subsystem_counts"], dataset["n_confs"])
                 ),
@@ -174,7 +325,7 @@ class TorchDataset(torch.utils.data.Dataset[Dict[str, torch.Tensor]]):
         for key, val in value.items():
             self.properties_of_interest[key][idx] = val
 
-    def __getitem__(self, idx: int) -> Dict[str, torch.Tensor]:
+    def __getitem__(self, idx: int) -> BatchData:
         """
         Fetch a dictionary of the values for the properties of interest for a given conformer index.
 
@@ -185,17 +336,7 @@ class TorchDataset(torch.utils.data.Dataset[Dict[str, torch.Tensor]]):
 
         Returns
         -------
-        Dictionary, contains:
-            - atomic_numbers: torch.Tensor, shape [n_atoms]
-                Atomic numbers for each atom in the molecule.
-            - positions: torch.Tensor, shape [n_atoms, 3]
-                Coordinates for each atom in the molecule.
-            - E: torch.Tensor, shape []
-                Scalar energy value for the molecule.
-            - idx: int
-                Index of the conformer in the dataset.
-            - atomic_subsystem_counts: torch.Tensor, shape [1]
-                Number of atoms in the conformer. Length one if __getitem__ is called with a single index, length batch_size if collate_conformers is used with DataLoader
+        BatchData instance representing the data for one conformer.
         """
         series_atom_start_idx = self.series_atom_start_idxs_by_conf[idx]
         series_atom_end_idx = self.series_atom_start_idxs_by_conf[idx + 1]
@@ -209,18 +350,26 @@ class TorchDataset(torch.utils.data.Dataset[Dict[str, torch.Tensor]]):
         ]
         E = self.properties_of_interest["E"][idx]
         F = self.properties_of_interest["F"][series_atom_start_idx:series_atom_end_idx]
+        total_charge = self.properties_of_interest["Q"][idx]
+        number_of_atoms = len(atomic_numbers)
 
-        Q = self.properties_of_interest["Q"][idx]
+        nnp_input = NNPInput(
+            atomic_numbers=atomic_numbers,
+            positions=positions,
+            total_charge=total_charge,
+            atomic_subsystem_indices=torch.zeros(number_of_atoms, dtype=torch.int32)
+        )
 
-        return {
-            "atomic_numbers": atomic_numbers,
-            "positions": positions,
-            "total_charge": Q,
-            "E": E,
-            "F": F,
-            "atomic_subsystem_counts": torch.tensor([atomic_numbers.shape[0]]),
-            "idx": idx,
-        }
+        metadata = Metadata(
+            E=E,
+            F=F,
+            atomic_subsystem_counts=torch.tensor([number_of_atoms], dtype=torch.int32),
+            atomic_subsystem_indices_referencing_dataset=torch.repeat_interleave(torch.tensor([idx], dtype=torch.int32),
+                                                                                 number_of_atoms),
+            number_of_atoms=number_of_atoms
+        )
+
+        return BatchData(nnp_input, metadata)
 
 
 class HDF5Dataset:
@@ -587,7 +736,7 @@ class HDF5Dataset:
             )
 
     def _to_file_cache(
-        self,
+            self,
     ) -> None:
         """
                 Save processed data to a numpy (.npz) file.
@@ -707,7 +856,7 @@ class DatasetFactory:
 
     @staticmethod
     def create_dataset(
-        data: HDF5Dataset,
+            data: HDF5Dataset,
     ) -> TorchDataset:
         """
         Creates a TorchDataset from an HDF5Dataset, applying optional transformations.
@@ -854,7 +1003,7 @@ class DataModule(pl.LightningDataModule):
         return DatasetFactory().create_dataset(dataset)
 
     def _process_self_energies(
-        self, torch_dataset: TorchDataset, atomic_self_energies: "AtomicSelfEnergies"
+            self, torch_dataset: TorchDataset, atomic_self_energies: "AtomicSelfEnergies"
     ) -> None:
         """Calculate and subtract self energies from the dataset."""
         from modelforge.potential.processing import AtomicSelfEnergies
@@ -1037,45 +1186,43 @@ class DataModule(pl.LightningDataModule):
 from typing import Tuple
 
 
-def collate_conformers(conf_list: List[Dict[str, torch.Tensor]]) -> "BatchData":
-    """Concatenate the Z, R, and E tensors from a list of molecules into a single tensor each, and return a new
-    dictionary with the concatenated tensors."""
-    from modelforge.potential.utils import BatchData, Metadata, NNPInput
-
-    Z_list = []  # nuclear charges/atomic numbers
-    R_list = []  # positions
-    F_list = []  # forces
+def collate_conformers(conf_list: List[BatchData]) -> BatchData:
+    """Collate a list of BatchData instances with one conformer each into a single BatchData instance."""
+    atomic_numbers_list = []
+    positions_list = []
+    total_charge_list = []
     E_list = []  # total energy
-    Q_list = []  # total charge
-    atomic_subsystem_counts = []
-    conf_idxs = []
-    for idx, conf in enumerate(conf_list):
-        Z_list.append(conf["atomic_numbers"])
-        R_list.append(conf["positions"])
-        E_list.append(conf["E"])
-        F_list.append(conf["F"])
-        Q_list.append(conf["total_charge"])
-        atomic_subsystem_counts.extend(conf["atomic_subsystem_counts"])
-        conf_idxs.append(conf["idx"])
-    atomic_subsystem_counts_tensor = torch.tensor(atomic_subsystem_counts, dtype=torch.int32)
-    atomic_subsystem_indices = torch.repeat_interleave(torch.arange(len(conf_list)), atomic_subsystem_counts_tensor)
-    atomic_subsystem_indices_referencing_dataset = torch.tensor(conf_idxs, dtype=torch.int32)[atomic_subsystem_indices]
-    atomic_numbers_cat = torch.cat(Z_list)
-    total_charge_cat = torch.cat(Q_list)
-    positions_cat = torch.cat(R_list).requires_grad_(True)
-    F_cat = torch.cat(F_list).to(torch.float64)
-    E_stack = torch.stack(E_list)
+    F_list = []  # forces
+    atomic_subsystem_counts_list = []
+    atomic_subsystem_indices_referencing_dataset_list = []
+    for conf in conf_list:
+        atomic_numbers_list.append(conf.nnp_input.atomic_numbers)
+        positions_list.append(conf.nnp_input.positions)
+        total_charge_list.append(conf.nnp_input.total_charge)
+        E_list.append(conf.metadata.E)
+        F_list.append(conf.metadata.F)
+        atomic_subsystem_counts_list.append(conf.metadata.atomic_subsystem_counts)
+        atomic_subsystem_indices_referencing_dataset_list.append(
+            conf.metadata.atomic_subsystem_indices_referencing_dataset)
+    atomic_subsystem_counts = torch.cat(atomic_subsystem_counts_list)
+    atomic_subsystem_indices = torch.repeat_interleave(torch.arange(len(conf_list)), atomic_subsystem_counts)
+    atomic_subsystem_indices_referencing_dataset = torch.cat(atomic_subsystem_indices_referencing_dataset_list)
+    atomic_numbers = torch.cat(atomic_numbers_list)
+    total_charge = torch.cat(total_charge_list)
+    positions = torch.cat(positions_list).requires_grad_(True)
+    F = torch.cat(F_list).to(torch.float64)
+    E = torch.stack(E_list)
     nnp_input = NNPInput(
-        atomic_numbers=atomic_numbers_cat,
-        positions=positions_cat,
-        total_charge=total_charge_cat,
+        atomic_numbers=atomic_numbers,
+        positions=positions,
+        total_charge=total_charge,
         atomic_subsystem_indices=atomic_subsystem_indices
     )
     metadata = Metadata(
-        E=E_stack,
-        F=F_cat,
-        atomic_subsystem_counts=atomic_subsystem_counts_tensor,
+        E=E,
+        F=F,
+        atomic_subsystem_counts=atomic_subsystem_counts,
         atomic_subsystem_indices_referencing_dataset=atomic_subsystem_indices_referencing_dataset,
-        number_of_atoms=atomic_numbers_cat.numel(),
+        number_of_atoms=atomic_numbers.numel(),
     )
     return BatchData(nnp_input, metadata)
