+++ modelforge/potential/utils.py
@@ -25,160 +25,6 @@ class NeuralNetworkData:
 from typing import NamedTuple
 
 
-@dataclass
-class NNPInput:
-    """
-    A dataclass to structure the inputs for neural network potentials.
-
-    Attributes
-    ----------
-    atomic_numbers : torch.Tensor
-        A 1D tensor containing atomic numbers for each atom in the system(s).
-        Shape: [num_atoms], where `num_atoms` is the total number of atoms across all systems.
-    positions : torch.Tensor
-        A 2D tensor of shape [num_atoms, 3], representing the XYZ coordinates of each atom.
-    atomic_subsystem_indices : torch.Tensor
-        A 1D tensor mapping each atom to its respective subsystem or molecule.
-        This allows for calculations involving multiple molecules or subsystems within the same batch.
-        Shape: [num_atoms].
-    total_charge : torch.Tensor
-        A tensor with the total charge of molecule.
-        Shape: [num_systems], where `num_systems` is the number of molecules.
-    """
-
-    atomic_numbers: torch.Tensor
-    positions: Union[torch.Tensor, Quantity]
-    atomic_subsystem_indices: torch.Tensor
-    total_charge: torch.Tensor
-
-    def to(
-        self,
-        *,
-        device: Optional[torch.device] = None,
-        dtype: Optional[torch.dtype] = None,
-    ):
-        """Move all tensors in this instance to the specified device/dtype."""
-
-        if device:
-            self.atomic_numbers = self.atomic_numbers.to(device)
-            self.positions = self.positions.to(device)
-            self.atomic_subsystem_indices = self.atomic_subsystem_indices.to(device)
-            self.total_charge = self.total_charge.to(device)
-        if dtype:
-            self.positions = self.positions.to(dtype)
-        return self
-
-    def __post_init__(self):
-        # Set dtype and convert units if necessary
-        self.atomic_numbers = self.atomic_numbers.to(torch.int32)
-        self.atomic_subsystem_indices = self.atomic_subsystem_indices.to(torch.int32)
-        self.total_charge = self.total_charge.to(torch.int32)
-
-        # Unit conversion for positions
-        if isinstance(self.positions, Quantity):
-            positions = self.positions.to(unit.nanometer).m
-            self.positions = torch.tensor(
-                positions, dtype=torch.float32, requires_grad=True
-            )
-
-        # Validate inputs
-        self._validate_inputs()
-
-    def _validate_inputs(self):
-        if self.atomic_numbers.dim() != 1:
-            raise ValueError("atomic_numbers must be a 1D tensor")
-        if self.positions.dim() != 2 or self.positions.size(1) != 3:
-            raise ValueError("positions must be a 2D tensor with shape [num_atoms, 3]")
-        if self.atomic_subsystem_indices.dim() != 1:
-            raise ValueError("atomic_subsystem_indices must be a 1D tensor")
-        if self.total_charge.dim() != 1:
-            raise ValueError("total_charge must be a 1D tensor")
-
-        # Optionally, check that the lengths match if required
-        if len(self.positions) != len(self.atomic_numbers):
-            raise ValueError(
-                "The size of atomic_numbers and the first dimension of positions must match"
-            )
-        if len(self.positions) != len(self.atomic_subsystem_indices):
-            raise ValueError(
-                "The size of atomic_subsystem_indices and the first dimension of positions must match"
-            )
-
-    def as_namedtuple(self) -> NamedTuple:
-        """Export the dataclass fields and values as a named tuple."""
-
-        import collections
-        from dataclasses import dataclass, fields
-
-        NNPInputTuple = collections.namedtuple(
-            "NNPInputTuple", [field.name for field in fields(self)]
-        )
-        return NNPInputTuple(*[getattr(self, field.name) for field in fields(self)])
-
-    def as_jax_namedtuple(self) -> NamedTuple:
-        """Export the dataclass fields and values as a named tuple.
-        Convert pytorch tensors to jax arrays."""
-
-        from dataclasses import dataclass, fields
-        import collections
-        from pytorch2jax.pytorch2jax import convert_to_jax
-
-        NNPInputTuple = collections.namedtuple(
-            "NNPInputTuple", [field.name for field in fields(self)]
-        )
-        return NNPInputTuple(
-            *[convert_to_jax(getattr(self, field.name)) for field in fields(self)]
-        )
-
-
-import torch
-
-
-@dataclass(frozen=False)
-class Metadata:
-    """
-    A NamedTuple to structure the inputs for neural network potentials.
-
-    Parameters
-    ----------
-    """
-
-    E: torch.Tensor
-    atomic_subsystem_counts: torch.Tensor
-    atomic_subsystem_indices_referencing_dataset: torch.Tensor
-    number_of_atoms: int
-    F: torch.Tensor = torch.tensor([], dtype=torch.float32)
-
-    def to(
-        self, device: Optional[torch.device] = None, dtype: Optional[torch.dtype] = None
-    ):
-        """Move all tensors in this instance to the specified device."""
-        if device:
-            self.E = self.E.to(device)
-            self.F = self.F.to(device)
-            self.atomic_subsystem_counts = self.atomic_subsystem_counts.to(device)
-            self.atomic_subsystem_indices_referencing_dataset = (
-                self.atomic_subsystem_indices_referencing_dataset.to(device)
-            )
-        if dtype:
-            self.E = self.E.to(dtype)
-            self.F = self.F.to(dtype)
-        return self
-
-
-@dataclass
-class BatchData:
-    nnp_input: NNPInput
-    metadata: Metadata
-
-    def to(
-        self,
-        device: Optional[torch.device] = None,
-        dtype: Optional[torch.dtype] = None,
-    ):
-        self.nnp_input = self.nnp_input.to(device=device, dtype=dtype)
-        self.metadata = self.metadata.to(device=device, dtype=dtype)
-        return self
 
 
 def shared_config_prior():
