+++ modelforge/tests/test_spk.py
@@ -140,7 +140,7 @@ def setup_single_methane_input():
     )
     E = torch.tensor([0.0], requires_grad=True)
     atomic_subsystem_indices = torch.tensor([0, 0, 0, 0, 0], dtype=torch.int32)
-    from modelforge.potential.utils import NNPInput
+    from modelforge.dataset.dataset import NNPInput
 
     modelforge_methane = NNPInput(
         atomic_numbers=atomic_numbers,
