+++ modelforge/train/training.py
@@ -4,7 +4,7 @@ import lightning as pl
 from typing import TYPE_CHECKING, Any, Union, Dict, NamedTuple, Tuple, Type, Mapping
 import torch
 from loguru import logger as log
-from modelforge.potential.utils import BatchData, NNPInput
+from modelforge.dataset.dataset import BatchData, NNPInput
 from torch.nn import functional as F
 
 if TYPE_CHECKING:
